refactor(dsg): change input type to get null
