+++ packages/amplication-data-service-generator/src/server/resource/dto/create-field-class-property.ts
@@ -8,11 +8,7 @@ import {
 } from "prisma-schema-dsl";
 import { Entity, EntityField } from "../../../types";
 import { classProperty, createGenericArray } from "../../../util/ast";
-import {
-  isEnumField,
-  isRelationField,
-  isToManyRelationField,
-} from "../../../util/field";
+import { isEnumField, isToManyRelationField } from "../../../util/field";
 import {
   createEnumName,
   createPrismaFields,
@@ -32,19 +28,18 @@ import {
 } from "./class-validator.util";
 import { INPUT_JSON_VALUE_KEY } from "./constants";
 import { createEnumMembers } from "./create-enum-dto";
+import { createGraphQLFieldDecorator } from "./create-graphql-field-decorator";
 import { InputTypeEnum } from "./create-input";
 import { createWhereUniqueInputID } from "./create-where-unique-input";
 import {
   EnumScalarFiltersTypes,
   SCALAR_FILTER_TO_MODULE_AND_TYPE,
 } from "./filters.util";
-import { GRAPHQL_JSON_OBJECT_ID } from "./graphql-type-json.util";
-import { FIELD_ID } from "./nestjs-graphql.util";
 import { createCreateNestedManyWithoutInputID } from "./to-many/create-create-nested-many-without-input";
 import { createUpdateManyWithoutInputID } from "./to-many/create-update-many-without-input";
 import { JSON_VALUE_ID } from "./type-fest.util";
 
-const DATE_ID = builders.identifier("Date");
+export const DATE_ID = builders.identifier("Date");
 const PRISMA_SCALAR_TO_TYPE: {
   [scalar in ScalarType]: TSTypeKind;
 } = {
@@ -149,12 +144,12 @@ export function createFieldClassProperty(
   optional: boolean,
   isQuery: boolean,
   isObjectType = false,
-  inputType: InputTypeEnum
+  inputType: InputTypeEnum | null
 ): namedTypes.ClassProperty {
   const [prismaField] = createPrismaFields(field, entity);
   const id = builders.identifier(field.name);
   const isEnum = isEnumField(field);
-  const isInput = !(inputType === InputTypeEnum.NotInput);
+  const isInput = inputType ? true : false;
   const [type, arrayElementType] = createFieldValueTypeFromPrismaField(
     entity.pluralDisplayName,
     field,
@@ -270,120 +265,6 @@ export function createFieldClassProperty(
   );
 }
 
-export function createGraphQLFieldDecorator(
-  prismaField: ScalarField | ObjectField,
-  isEnum: boolean,
-  field: EntityField,
-  optional: boolean,
-  entity: Entity,
-  isQuery: boolean,
-  inputType: InputTypeEnum,
-  isNestedInput: boolean
-): namedTypes.Decorator {
-  const type = builders.arrowFunctionExpression(
-    [],
-    createGraphQLFieldType(
-      prismaField,
-      field,
-      isEnum,
-      entity,
-      isQuery,
-      inputType,
-      entity.pluralDisplayName,
-      isNestedInput
-    )
-  );
-  return builders.decorator(
-    builders.callExpression(
-      FIELD_ID,
-      optional || isQuery || !field.required
-        ? [
-            type,
-            builders.objectExpression([
-              builders.objectProperty(NULLABLE_ID, TRUE_LITERAL),
-            ]),
-          ]
-        : [type]
-    )
-  );
-}
-
-function createGraphQLFieldType(
-  prismaField: ScalarField | ObjectField,
-  field: EntityField,
-  isEnum: boolean,
-  entity: Entity,
-  isQuery: boolean,
-  inputType: InputTypeEnum,
-  entityPluralName: string,
-  isNestedInput: boolean
-): namedTypes.Identifier | namedTypes.ArrayExpression {
-  if (prismaField.isList && (!isToManyRelationField(field) || isNestedInput)) {
-    const itemType = createGraphQLFieldType(
-      { ...prismaField, isList: false },
-      field,
-      isEnum,
-      entity,
-      isQuery,
-      inputType,
-      entityPluralName,
-      isNestedInput
-    );
-    return builders.arrayExpression([itemType]);
-  }
-  if (isQuery && prismaField.kind === FieldKind.Scalar) {
-    return getFilterASTIdentifier(field.required, prismaField.type);
-  }
-
-  if (prismaField.type === ScalarType.Boolean) {
-    return BOOLEAN_ID;
-  }
-  if (prismaField.type === ScalarType.DateTime) {
-    return DATE_ID;
-  }
-  if (
-    prismaField.type === ScalarType.Float ||
-    prismaField.type === ScalarType.Int
-  ) {
-    return NUMBER_ID;
-  }
-  if (prismaField.type === ScalarType.String) {
-    return STRING_ID;
-  }
-  if (prismaField.type === ScalarType.Json) {
-    return GRAPHQL_JSON_OBJECT_ID;
-  }
-  if (isEnum) {
-    const enumId = builders.identifier(createEnumName(field, entity));
-    return enumId;
-  }
-  if (isRelationField(field)) {
-    if (isNestedInput) {
-      return createWhereUniqueInputID(prismaField.type);
-    }
-    if (isToManyRelationField(field)) {
-      switch (inputType) {
-        case InputTypeEnum.Create:
-          return createCreateNestedManyWithoutInputID(
-            entityPluralName,
-            field.properties.relatedEntity.name
-          );
-
-        case InputTypeEnum.Update:
-          return createUpdateManyWithoutInputID(
-            entityPluralName,
-            field.properties.relatedEntity.name
-          );
-        default:
-          throw new Error("Didnt got an input type");
-      }
-    }
-    return createWhereUniqueInputID(prismaField.type);
-  }
-
-  throw new Error("Could not create GraphQL Field type");
-}
-
 export function createTypeDecorator(
   typeName: namedTypes.Identifier
 ): namedTypes.Decorator {
@@ -406,10 +287,8 @@ export function createFieldValueTypeFromPrismaField(
   isQuery: boolean,
   isObjectType: boolean,
   isNestedInput: boolean,
-  inputType: InputTypeEnum
+  inputType: InputTypeEnum | null
 ): TSTypeKind[] {
-  const isInput = !(inputType === InputTypeEnum.NotInput);
-
   // add  "| null" to the end of the type
   if (
     !prismaField.isRequired &&
@@ -498,7 +377,7 @@ export function createFieldValueTypeFromPrismaField(
       ),
     ];
   }
-  if (isQuery || isInput || isNestedInput) {
+  if (isQuery || inputType || isNestedInput) {
     return [
       builders.tsTypeReference(createWhereUniqueInputID(prismaField.type)),
     ];
@@ -507,7 +386,7 @@ export function createFieldValueTypeFromPrismaField(
   }
 }
 
-function getFilterASTIdentifier(
+export function getFilterASTIdentifier(
   isRequired: boolean,
   type: ScalarType
 ): namedTypes.Identifier {
