+++ packages/amplication-data-service-generator/src/version.ts
@@ -1,1 +1,1 @@
-export const version = "0.11.2";
+export const version = "0.11.3";
